Basic Control Plane Creation (#6)
Teeing up for tomorrow.  This accepts and validates a control-plane name
and creates a namespace for it.  The namespace is labelled with an "I'm
a control plane" label which will allow `unikorn get control-plane`
functionality to be written along with delete, that will need to
validate it's not deleting `kube-system` :smile_cat:.