Clean Up Get Projects (#10)
There was a lot of crap being pulled in from kubectl that's not
necessary e.g label selectors and pointless, field selectors are useless
etc.  This removes them so only the core functionality is exposed.  I've
also taken the liberty to add in completion for the output format flag,
and gotten around to coalescing multiple return arguments into a single
list as is the expected behaviour of kubectl.