Add Kind LoadBalancer Support (#12)
To aid local and CI testing, we need to provide load balancer services
in order to use vcluster.  To that end, we provide a hack script that
installs MetalLB and provides LB service support with docker
integration.  There're quite a few utility functions that will come in
handy elsewhere when it comes to 'wierd' provisioning strategies, and
tests for that matter, will refactor them when necessary.