Provision Vcluster (#16)
Admittedly badly, given we cannot propagate kubeconfig and context
parameters.  But!  We can provision it into the project namespace with
the correct name for the control plane and wait for readiness.  Now it
remains to be seen whether the kubeconfig is ready at that point or it's
an after thought...