Add in More Control Plane Boilerplate (#17)
Deletion being the operative part here.  Now, we alredy know
installation via a binary is terrible.  Deletion more so as it will
leave a bunch of crap lying around, and given it all came from a black
box box we cannot clean it up ourselves without having to resort to
doing a deprovisioning step.  This leads to me thinking we need to run
helm template, modify the YAML (JSON) to allow cascading deletes and
any other manipulation we see fit.  It also allows us full control of
the install via a dynamic client...