Vcluster Tweaks, and Fixes
Fix a bug in the CRD printcolumns introduced by linting.  "Fix" a
persistence thing with vcluster that is causing stuff to raise from the
dead like necromancy.  Add in a tweak to ensure CAPI can provision in
light of OpenStack being slooooow.  Finally add in a command to get the
vcluster Kubernetes config for debug purposes.