FIrst Crack at Prometheus (#27)
We're _passionate_ about performance, so having metrics is a good thing.
As a start I've augmented the vcluster provisioner to see what it looks
like.  Turns out there was a whole bunch of noise in there, so I've
tweaked the controllers to ignore status updates, which should have been
done to start with, but it did test idenpotency.  Also took the
opportunity to tweak the status reporting code now it's actually sane,
and add in some configurable timeouts.