Apple Silicon Fixes (#31)
Updates README with note on using GNU/Make.  

If you build images on a non-x86 machine you'll end up with container
images for potentially the wrong target platform.  Update the Makefile
to use docker buildx and explicitly specify x86 as the target.

As we are using buildx, seems it doesn't need the buildkit envvar any
more, so bin that.  Try as I might to get a working MacOS CI job, screw
them all essentailly.  I did think it prudent to use a bit of code
sharing though.

Co-authored-by: Nick Jones <nick@dischord.org>