Add in Some More Metrics (#32)
Toyed with making this a struct, then decided against it as propagation
is going to be a nightmare.  Adds in timers for cluster API and a global
one around control plane provisioning.  Ran this locally in Kind to get
a vague understanding of how log we should expect and what the buckets
should be setup like.  In the comments for each histogram.