Cluster Provisioning Part 1 (#37)
Add in a starter-for-ten CRD that represents a basic Kubernetes cluster.
Start fleshing out the unikornctl command to actually create CRs.  For
now it's able to grab a clouds.yaml, extract the specific one you care
about, then also derive the CA certificate by dialing the public
endpoint, that'll do a verification, at which point we can pluck out the
CA that verified the server chain.