Cluster Provisioning Part 4
Adds in a cluster manager to provision cluster resources.  Few CRD and
tooling tweaks based on what the actual provisioner requires.  The main
change here is adding JSON patching to manifests so we can override
parts of the vanilla vendor manifests e.g. allowing fine grained control
of networking in this case.