Make Cluster Provisioning Work
The main fix here was to have vcluster clients reinitialise their
dynamic clients against the target cluster, so it picks up the CRDs
installed there.  Unbeknownst to me, it will also refresh the REST
mapper when it encounters something unknown, which is pretty awesome.
Other than that, remove a broken image repository from the cluster
manifest and fix the cluster provider configuration for kubelet.