Add Cluster Get to Unikornctl
First up we move the provisioning of clusters into a namespace, this
gives us a nice primitive to delete.  Next we remove the control plane
from the CR and make it a required label, that way we can filter the
cluster by control plane.  Cleaned up the printing code for better code
reuse (it was getting a cyclo error - sigh).  Finally make an idempotent
provisioner for single resources e.g. that namespace ^^.  I could
probably make the manifest provisioner call the resource provisioner for
each YAML resource...