Fix Cluster Deletion
Because CAPI is a bit broken, we need to ensure Openstack secrets are
kept alive for the duration of cluster deprovision, or it won't be able
to talk to Openstack, and sit there forever waiting on finalizers to
complete that cannot.  It's a bit of a hack, but we can filter out
secrets from the deprovision of the manifest.  Once CAPI has done its
thing, the resources will be deleted, the cluster manager unblocked and
the namespace can be deleted to finish off the cleanup.