Add in CNI Provisioning
While the cluster is provisioning we need to, in parallel, create the
remote cluster in ArgoCD and then install Cilium so the cluster
application eventually becomes healthy.  The hope here is CoreDNS won't
come up because CAPI doesn't provision a CNI, so we don't have to mess
around with setting up taints.