Expose Workload Pools
Further to the last commit, propagate worklaod pools up the stack so
they are exposed at the API level.  In order for this to ever work with
CLI tooling wee need to have a new workload pool custom resource.  There
are a few rough edges that need fixing e.g. detecting changes to
workload pools in the controller, and clean up like applying owner
references, but that comes next.