First Stab at Autoscaling (#75)
Adds a separate get for kubeconfigs, as the actual cluster has been
overlooked for too long.  As a result, the vcluster stuff needed a heap
of love to make it sane.  Finally add in an initial cut of autoscaler
provision/deprovision.  Additionally discovered a bug with GPU labels,
so bumping the helm revision for that.