Autoscaling Fixes (#91)
Remove namepsace scoped RBAC for the autoscaler as that's broken, and
ignore the replicas field in machine deployments because cluster
autoscaler updates the spec when it shouldn't. Additionally based on
experience/feedback, remove the region as it's populated by the cloud
controller or cloud config, make ssh keys optional, remove the default
workload pool hack, update all the cluster api charts to pick up fixes
and enable the above.  Finally fix an idempotency bug in the
deprovisioning of clusters.