Make Remote Handling More Generic (#95)
Refactor out destination knowledge from the application Genrerate
functions and make it a genric function of the application itself.  In
doing so we can propagate the remote generator down and benefit from
whatever that gives us in the future.  Undid a couple changes to do with
propagation of vcluster clients, it's adding in unnecessary coupling
that looks ugly.