Autoscale Fixes and Scale to Zero (#100)
Use a new cluster helm chart that will not set the replicas at all when
autoscaling is enabled, it transpires that CAPI erroneously sets the MD
replicas when the scale subresource is updated, this collides with Argo
which then reverts the change.  Classic split brain that they won't fix.
Update the scheduling and toleratons for the nVidia GPU operator to move
it off the workload pools, and therefore can scale to zero.