Use forked cluster-autoscaler Helm chart for SFZ support
For scale-from-zero to work, the Service Account needs some additional
provider-specific permissions.  See
https://github.com/kubernetes/autoscaler/blob/master/cluster-autoscaler/cloudprovider/clusterapi/README.md#rbac-changes-for-scaling-from-zero
for the background and details.

This PR updates the application definition to use our fork which
includes the necessary patch for CAPO.

Additionally, SFZ support was added as of v1.25 whereas the Helm chart
default is still to deploy v1.23.  We explicitly install the latest
release version instead by specifying the tag.