OpenAPI Request/Result Validation
First up add some more static validation to ensure our policies are
being enforced with API design/security.  Add in validation so we can
reject requestes outright that fail our validation, and warn when we
cock up results.  The OAPI validation also delegates authn/authz to a
callback, so we may as well use that rather than maintaining code and
using a noop.