Control Plane Firewall & Friends
The control plane needs access to the managed Kubernetes instance e.g.
to drain etc.  As such it needs to penetrate the firewall.  To make
things usable we "can haz ip" and implicitly add that to the allow list.
Additionally there was a bunch of missing logging that allowed me to
track down a hang in deprovisioning due to the NVIDIA license hack.
Finally roll back the server VGPU crap, it's over engineering.