Complete Project API
This was overly complicated as the JWT implies the project, so make that
available in the claims to the handler.  Second, remove parameterized
projects from the APIs.  They may be added back in later if you have an
admin scope, but it's not necessary for now, keep it simple, keep it
secure.  Create a global client for caching purposes, although I'm not
totally sure it does as it requires get perms, so needs more
investigation.  Add in the projects CRUD interface, and also takes the
liberty of updating the OAPI spec across the board to save effort later.