Add Application Credential Interface
Clusters a provisioned scoped to a specific project using application
credentials.  Not only does this provide least-priviliege (passwords can
be scoped to multiple projects), it's also immutable.  This provides a
way to check for a named credential, create a new one (needs to be
per-cluster as we only see the secret once), and delete them.  While we
are here, setting up configuration or roles, rearchitect the whole flags
part, using 3rd party libraries to reduce code.