Complete Cluster API (#126)
First up, this isn't the CLI, so we don't need to create workload pools
piecemeal, we can do it in one shot, so allow them to be specified
inline in the custom resource. Other than that, standard stuff, like the
projects and control planes before them.  Managed to refactor some
common stuff between the CLI and server.