UI Charts and Friends!
Adds in the UI component to the server chart as they are linked.  Tweaks
the token code to propagate all claims up to the handlers so you can
return things like email addresses to a client, rather than just an
opaque token!  Tries to add in cache control headers, but they don't
seem to have an effect on firefox at least, perhaps there's something
wrong with the CA, perhaps because it's manually installed.