Improve OTEL
Extend the OTEL provider to send traces via OTLP to something like
Jaeger so we can see where the time is spent visually.  To aid in this a
lot of the Openstack provider code has been augmented to allow the APIs
to be instrumented.  Several componenents have been refactored to yield
cleaner code.  Documentation, Helm, blah, blah, blah.