Handle annotations for server Ingress resource
Allow specifying annotations for the Ingress resource that's created for
use by the Server component.  Useful if, for example, someone wants to
specify the use of a ClusterIssuer that's defined elsewhere in order to
provide TLS certificate issuance, or if additional annotations are
required for components such as ExternalDNS.

Remove the ACME configuration but leave in place a default Issuer for
development and testing.