Bunp Cluster Chart
Fixes a race condition we trigger with CAPO deleting its machines and
infrastructure at the same time.  In this case the openstackcluster gets
killed and CAPO cannot deprovision the machines.  What this does is
provisions cluster and MDs last, so they get deleted first, and in
theory this problem doesn't happen.  However Argo is broken and appears
to schedule a refecsh during delete, so to mitigate, turn of auto sync
on deletion.