Auto Upgrade Part 1 (#184)
Adds in a new unikorn-monitor component that sits in a loop and
periodcally calls a modular set of plugins.  The two that are there
handle application bundle upgrades for CP and CL resources... ish.  I've
added enough bits to check that upgrade paths work as expected, and also
that opt-in does its job.  Going to pivot slightly to do the UI work
while I muse over upgrade time windows and how best to describe those in
Kubernetes resources.