Add Support for Explicit Time Windows
Most humans (and this is based on experience, not just conjecture), like
to have the ability to specify what days things run, and during what
timeframe.  So we provide an auto-upgrade interface that allows you to
do this.  I've taken the liberty to allow windows to wrap, so you can go
Sun 23:00-07:00, leaving Monday free, rather than forcing you into Sun
23:00-00:00 & Mon 00:00-07:00, then having to support multiple windows.
Like implicit auto-upgrade, that window is a hint, we use a similar
scheduling algorithm within the user window to provide load balancing.
Onwards we go with forced upgrades!