Scale-From-Zero Fixes
More precisely scale-to-zero.  We'd always have one node, because a
bunch of managed stuff wasn't pinned to the control plane e.g. cilium
operator, cinder CSI, and most importantly coreDNS.  Now on 1.26 the
"master" label has been removed because of cancel culture, so caoreDNS
only tolerates the "control-plane" label and cannot be scheduled there
due to the legacy "master" stuff.  So we've updated the cluster chart to
not taint control plane nodes with this label.  Now, CAPI being CAPI
could choose to run a kubectl command against the API to reconcile this,
but instead deems it appropriate to do a rolling upgrade instead!  So be
warned.  Also, we need to get rid of the nvidia licence override, as
it's baked into the image.  At present, there's no easy way to remove
this as it's outside of Argo's remit, so just leave it in place, and
it'll vanish on any for of upgrade over time... just be aware.