Finish Monitor Upgrades (#189)
It came to me in a dream... that auto upgrades are just a subclass of
day of the week upgrades, so refactor that to use the same code.
Slightly modify the algorithm as a result.  Add in forced upgrades when
the bundle is EOL and we're after the time limit.