Fix Namespace Scoping (#198)
The namespace lookup was first written with projects, these are global
and have unique names, fine.  What we missed was controlpanes are
scoped, so you can reuse the name in different projects, so going by
name was totally broken.  Like argo apps, use the resource labels
function to define a unique selector set, so we rely on control plane
name and the project name to provide scoping.  However this in itself is
broken in that control planes namespaces now have a project label,
therefore will get picked up by the project controller, so we need a new
kind label to tell them apart.  But this is broken in that old
namespaces have none of the new labels, so we need an upgrade mechanism
to add them.