Workaround for ArgoCD's insistence on mangling quoted bools
The NVIDIA Operator's ClusterPolicy validation explicitly expects a
quoted boolean, however along the way ArgoCD flips this from as such to
a bool.  One option is to set forceString to true in the Application CR,
but as we don't expose this via Unikorn this commit is the workaround
where it's ostensibly a string but is correctly interpreted by the
NVIDIA Operator.

Refs:

https://github.com/NVIDIA/gpu-operator/blob/9892f06307a2498c2e6eeed43612479919d89801/validator/main.go#L244
https://github.com/NVIDIA/gpu-operator/blob/9892f06307a2498c2e6eeed43612479919d89801/vendor/github.com/urfave/cli/v2/flag_bool.go#L65
https://github.com/argoproj/argo-cd/issues/4428