Add Kubernetes Dashboard (#211)
Standard fayre... with the exception of the DNS/TLS shizzle, as that
requires us to know in advance the DNS name to inject into the
certificate.  Usually we'd use external-dns, but that's not very secure
as it means a private token in the user's cluster. Instead we opt for
wildcard DNS, but that has its own problems in that we now need to know
the ingress IP address, which means a fairly intusive poking about.