Inhibit Non-Production Upgrades (#214)
So when you build non-prod, you end up using the default version 0.0.0,
which will trigger every upgrade under the sun.  Problem is whatever's
been deployed with this version will use the newest schema/set of
requirements, so the upgrades will fail.  Rather than trying to guess
what to do (which is to be frank, a hard problem), just opt out of
upgrades.

There is a danger that if any changes happen that require an upgrade,
things will go horribly wrong :D  But still it's less likely that the
above, which will happen on any restart after creating a resource.