Fix Autoscaler SNAFU (#218)
One for @yankcrime to be aware of.  When upgrading from an old resource
version to a new, and you add a new required chart (in this case the
autoscaler roles, required in the context of the autoscaler), then you
need to be aware that older version will not have this application
available.  It only exists in the new resource bundle.  As such it
should be made optional (temporarily), and conditional based on the
presence of the application.  See the code for more details.

Now I'm not Superman, so it slipped through review, but not integration
testing.  For future reference, what I do is start with an old unikorn
deployment, deploy some clusters in a variety of configurations (in this
case one with autoscaling, as I knew you touched it).  Do the unikorn
upgrade, and watch for errors, fix them, proceed to do other upgrade
tests.  One day we'll have automation... but that takes precious time.