Make Server Groups Work E2E (#237)
Fairly simple now the hard bit has been done, just adding in the CRD and
API fields.  Now, I've updated the cluster helm chart in place, as it's
a non-affecting change, so we'll need to wait until the end of the month
before upgrading everything due to the 1.1.0 cluster bundle to go full
EoL and all the clusters to upgrade to at least 1.2.0.  Unikorn UI will
ensure any new clusters are provisioned with server groups.