Workload Pool Deletion FIx
We delete workload pools manually based on the name generated by helm,
which is s***e.  So add in a temporary hack, and add in a new chart that
will augment the pool resources with a label so we can directly a) find
resources related to that pool and b) compare names natively without any
knowledge of the chart's naming/hashing nuances.