Move Application Credentials into Server
Having the client manage these is extra work for the client, plus the
application credential secret traverses the public internet twice.
Doing this as part of Server not only makes it more secure, but also
improves in the interface. Additionally sneaks in the nvidia operator
option so I can upgrade existing clusters before making the actual
change live.