Remove "all" Category (#263)
Part of the reason the old management cluster got totalled is because
our CRDs are maked as belonging in the "all" category.  We should remove
this to make it harder to delete things.  Besides listing all the
applications and bundles when doing "all" is kind of a pain in the butt.