Fix Credential Rotation
As per usual the "community" are resisting the correct change of having
the controllers pick up config changes automatically, instead insisting
that every client on the planet needs to manually restart things to
achieve that.  Sadly, it's impossible to do this unless you're on the
latest and greatest helm chart, so that needs upgrading, then as we need
to hash things, we may as well just define our own cloud-config and hash
that... hopefully for the time being.