Implicit Child Object Creation
At present, a client needs to create a project, wait, create a control
plane, wait, then create a cluster.  Theoretically it should be possible
to create a cluster, have it create a default control plane, wait only
as long as it takes to provision the namespace and yield, etc.

Theory and practice are very different though, with the main "simple"
bit done, it revealed that waiting is subtly flawed in that there is
always a delay before the first tick, so that needs fixing from a
performance standpoint.  Next there is a race where cluster creation
fails with an error due to CRD installation racing against machine
delployment cleanup.  This leads to the cluster getting stuck in an
error state.  This is now fixed so the error is transient, but does
require the core of all the managers to be ripped out and replaced.
Tracking down that issue in the first place required a whole heap of
extra debug code, which should have some benefits down the line, I had
been meaning to do something of that ilk.

Fixes #249