Extreme Openstack Provider API Testing (#296)
When creating a cluster, we will interrogate images and flavors to get
additional metadata bout them, as the APi only requires a name to
function only.  These calls can actually get an unauthorized error, and
creation of an application credential can get a forbidden, so ensure
these are correctly propagated to the client.