Test Application Provisioner (#300)
If ever there were a time for Nick to live in fear, it's now.  So it
transpires that the test clients cannot handle pointers to scalars when
doing a deep copy on unstructured data, so there were two options...
First hack it so pointers were checked and dereferenced, which looked
ugly as hell.  The other option was to define the Application data
structures and generate deep copy functions for it.  Obviously the
latter is the best approach, from a cleanliness and safety perspective.
However, as this is literally the core of the product, it's not without
a teeny bit of risk.  Hold on to your butts...