Refactor CD Layer
So the point here is that testing the remote cluster provisioner is
going to be pretty impossible as you are going to have to mock ArgoCD.
So what if we abstract away the CD bit, so we have a mockable interface?
Good news is that on top of separating the CD implementation from the
application and remote cluster provisioners, we can also, if needs
arise, swap out ArgoCD for Flux, which while isn't as flashy, will most
likely be more reliable from a resource constraint perspective, even if
it doesn't give guarantees about self-healing... swings and roundabouts.