Application Inference (#313)
I noted in the prior commit that handling of applications and bundles is
quite messy, and it's been brewing for quite some time, but all
applications have a name, and they should use that to lookup the
application from the avaialble bundle, rather than having the
application passed down to it.  Turns out it was relatively easy, and
does wonders for code clarity.  Quite surprisingly the stars mostly
aligned, and it was just the cluster-openstack that generated a
different Application name to its HelmApplication, but easily remedied.