Provisioner Option Inheritance (#315)
Take the cluster provisioner as an example, nearly all apps are being
installed on the remote cluster, and have background deletion enabled.
What would be infinitely cleaner is if we could specify these at the
level of an ancestory container and have them propagated to all
descendants.  Cleaner, easier to read.  Also, while in the area, I was
drawn to the resource provisioner, which is a bit archaic and plays
around with unstructured data types, still not perfect by any stretch of
the imagination, but baby steps!