Post Provision Application Hooks
Imbue the application provisioner with the power to invoke a
post-provision hook, which removes the last "special" helm application
that had to define its own provisioner type, rather than piggy-back on
the application.  This simplifies things, deletes a bunch of crap, and
provides a generic framework.  As the hook itself is an Argo specific
hack, provide a way so the hack knows what it's using and whether or not
action needs to be taken.