Fix CD Driver Initialisation
THis was hacked in and done every time we reconciled, but what we should
be doing is creating one driver and reusing it.  The other benefits are
that we can also do the driver selection bit and create a flag for it.
Hard part was integrating it with controller-runtime, but there is a way
to have it start after the client is up and running, and before
leadership election kicks in and unleashes the reconciler that will
require the driver to be available.