Propagate Stuff via the Context
The Kubernetes client, CD driver and owning resource are pretty much
propagated everywhere, which adds a lot of bloat to function signatures,
structs that have to cache these values etc. What would be a lot cleaner
is to pass these about in the context which is by necessity propagated
everywhere.  The added side benefit is we can override the values in
different scopes, for example, if we use a remote instance of Argo we
can change the client to the remote cluster, and the driver to the
remote Argo instance.