Make Remote Clusters Declarative
Fuuuu... there is a declarative way of setting this up in Argo, but not
done via a CRD (which is where I made the mistake), instead a secret.
So tear out all the API level stuff (which may possibly require complex
routing) and replace with something that generates, and most importantly
claims owneership of existing ones.