Implicit Client Scoping
Previously various bits and bobs needed to create clients and have
explicit knowledege of topology (criminal) and even implementation
details (even more criminal).  What we should do is update the clients
as the scope changes, so as we enter into a new remote cluster, the
client is also updated to interact with it.  This also handles nesting
which is even cooler.  The only real caveats here are we need a
distinction between a dynamic client (that changes as the scope does)
and a static one (that always points at the base cluster e.g. for access
to application bundles etc).  Also, this requires some better handing of
deprovisioning as we're now creating clients unconditionally when they
previously would have been lazily created.