Honour explicit requests to not install the NVIDIA GPU Operator
Right now, Unikorn Server's decision as to whether to install the
Operator or not is final as far as the resulting Kubernetes cluster spec
is concerned, based on whether GPUs are detected in any workload pools.

This change checks for an explicit disabling of said Operator in the
creation request, and if set does not install regardless of whether GPUs
are available.