Improve Cluster Start Up Timing
Argo will have a flap if you create a remote and it cannot talk to it,
stalling for like 5 minutes until it retires the conection, so perform
this check for it to avoid the timeout, much like how it would work if
created via the API.  Secondly, there's a quick hack to allow continued
provisioning while certain applications are degraded i.e. working, but
not fully.  Also, sneak in a change to CAPO to prevent it doing "clever"
scheduling choices, when Nova is actually more cleverer in this
instance.