Unconjoin Application Bundle Types (#344)
While having a single type for application bundles shared between CPs
and CLs is kinda nice, in that arlgorthms are shared, it leads to other
issues. You have to filter based on kind, which could lead to bugs, and
you cannot do type specific validation, for example you could choose to
enforce a set of applications for a specific kind.  This tries it out to
see what it could look like.