Refactor Application Handling (#358)
Prior to this change the application was aware only of a provisioning
resource type (CP or CL), and was responsible for grabbing the bundle,
then extracting the application.  This will not work with user
application bundles as it's already a bundle of sorts.  The common
denominator here is the application reference, thus the application
provisioner needs an interface to get an application reference from
whatever is provisioning the application.  Traspires this solution is
actually a fairly decent middle-ground.